% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_themes}
\alias{get_themes}
\title{Utility function to access predefined highcharts templates}
\usage{
get_themes(env)
}
\arguments{
\item{env}{a local environment created in each function and passed here to grab the
selected template}
}
\description{
Utility function to access predefined highcharts templates
}
