% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaBar.R
\name{rdaBar}
\alias{rdaBar}
\title{Highcharts horizontal bar chart}
\usage{
rdaBar(df, x, y, title = "", subtitle = "", tooltip_text = "", caption = "")
}
\arguments{
\item{df}{name of dataframe}

\item{x}{string of the independent variable to appear on y-axis}

\item{y}{string of the dependent variable to appear on x-axis}

\item{title}{provide as a string}

\item{subtitle}{provide as a string}

\item{tooltip_text}{provide as a string}

\item{caption}{provide as a string}
}
\value{
A responsive horinzontal bar chart
}
\description{
Highcharts horizontal bar chart
}
