% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaGroupedBar.R
\name{rdaGroupedBar}
\alias{rdaGroupedBar}
\title{Highchart horizontal grouped bar chart with RDA customization}
\usage{
rdaGroupedBar(
  df,
  x,
  y,
  z,
  bar_colors = list(meteorite, lavender, peridot, papaya, ccblue, black, gainsboro),
  theme = "theme_cc",
  title = "",
  subtitle = "",
  tooltip_text = "",
  caption = "",
  export_data_label = ""
)
}
\arguments{
\item{df}{Name of data frame with data to visualize}

\item{x}{String of the column name of the independent variable (e.g., Race/Ethnicity) to be plotted on the x-axis}

\item{y}{String of the column name of the dependent variable (e.g., Rates, Counts, etc.) to be plotted on the y-axis}

\item{z}{String of the column name of the grouping variable (e.g., subcategories to further cut the data within x) to create stacked bars}

\item{bar_colors}{List of hexcode colors that the chart should use (e.g., c("#EFEFEF", "#F25922", "#CEEA01"). By default will use pre-defined CC colors: meteorite, lavender, peridot, papaya, ccblue, black, gainsboro.}

\item{theme}{a hc_theme object. can be created by user or be a predefined project theme (options include: "theme_cc", "theme_fbhc"). default value is theme_cc.}

\item{title}{Chart title provided as string - appears at the top of chart}

\item{subtitle}{Chart subtitle provided as string (also name of exported file) - appears directly underneath title}

\item{tooltip_text}{Content displayed when cursor hovers over bars. Provide as string (can include Js and HTML syntax).}

\item{caption}{Chart caption provided as string - appears at bottom}

\item{export_data_label}{provide as a string (can include Js and HTML syntax), e.g., "    '{point.rate:.1f}\%'"}
}
\value{
An interactive grouped bar highchart
}
\description{
Highchart horizontal grouped bar chart with RDA customization
}
