% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{theme_options}
\alias{theme_options}
\title{An exposed list of predefined hc_themes}
\format{
A list with 2 elements.
}
\usage{
data(theme_options)
}
\description{
A list that can be accessed by users of the package.
}
\keyword{themes}
