% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaBar.R
\name{rdaBar}
\alias{rdaBar}
\title{Highcharts horizontal bar chart}
\usage{
rdaBar(
  df,
  x,
  y,
  theme = theme_cc,
  title = "",
  subtitle = "",
  tooltip_text = "",
  caption = "",
  export_data_label = ""
)
}
\arguments{
\item{df}{name of dataframe}

\item{x}{string of the independent variable to appear on y-axis}

\item{y}{string of the dependent variable to appear on x-axis}

\item{theme}{a hc_theme object. can be created by user or be a predefined project theme (options include: theme_cc, theme_fbhc). default value is theme_cc.}

\item{title}{provide as a string}

\item{subtitle}{provide as a string}

\item{tooltip_text}{provide as a string}

\item{caption}{provide as a string}

\item{export_data_label}{provide as a string (can include JS and HTML syntax), e.g., "    '{point.rate:.1f}\%'"}
}
\value{
A responsive horizontal bar chart
}
\description{
Highcharts horizontal bar chart
}
