% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaBubblepop.R
\name{rdaBubblepop}
\alias{rdaBubblepop}
\title{Highcharts Bubblepop chart with Catalyst California theme}
\usage{
rdaBubblepop(
  df,
  x,
  y,
  z,
  theme = "theme_cc",
  title = "",
  subtitle = "",
  tooltip_text = "",
  caption = "",
  yaxis_label = "''",
  export_data_label = ""
)
}
\arguments{
\item{df}{name of dataframe}

\item{x}{string of the independent variable to appear on y-axis}

\item{y}{string of the dependent variable to appear on x-axis}

\item{z}{string of the variable that stores count to determine bubble size}

\item{theme}{a hc_theme object. can be created by user or be a predefined project theme (options include: "theme_cc", "theme_fbhc"). default value is theme_cc.}

\item{title}{provide as a string}

\item{subtitle}{provide as a string}

\item{tooltip_text}{provide as a string}

\item{caption}{provide as a string}

\item{yaxis_label}{provide as a string}

\item{export_data_label}{provide as a string (can include Js and HTML syntax), e.g., "    '{point.rate:.1f}\%'"}
}
\value{
A responsive html bubblepop chart. Bubblepops are a combination of bar and
}
\description{
Highcharts Bubblepop chart with Catalyst California theme
}
